Updated test-eight.py
This still doesn't automatically strip out stuff outside of mainland Manhattan, nor does it strip out addresses that don't begin with a number (as I Quant NY appears to have done). But for demonstration purposes, it appears to get pretty close to their results, with 10 Gracie Square among the very last of the results when sorted by subway distance. Hurray!